Bill S2912 version original - Dean Murray
BILL NUMBER: S2912 SPONSOR: MURRAY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to yellow signals at traffic intersections PURPOSE: To standardize across the state a minimum amount of time for the yellow signal at a traffic intersection to stay lit prior to turning to red. SUMMARY OF PROVISIONS: Section one amends section 1111 of the vehicle and traffic law to provide that yellow traffic signals should provide a minimum of four seconds warning before turning red.