Bill S2929 version original - Gustavo Rivera
BILL NUMBER: S2929 SPONSOR: RIVERA TITLE OF BILL: An act to amend the public health law, in relation to the establishment, incorporation, construction, or increase in capacity of for-profit nurs- ing homes PURPOSE: Prohibits the establishment, incorporation, construction, or increase in capacity of for-profit nursing homes. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends the public health law by adding a new section 2802-c as it relates to expansion of for-profit nursing homes.