Bill S2906 version original - Dan Stec
BILL NUMBER: S2906 SPONSOR: STEC PURPOSE OR GENERAL IDEA OF BILL: SUMMARY OF PROVISIONS: Section 1: Calls upon the Office of Mental Health, Office of Alcoholism and Substance Abuse Services, and the Department of Education to create and maintain a website providing information on a number of behavioral health issues. This section also states that the website contains an extensive list of contact information and other resources for treatment facilities.