Bill S2916 version original - Cordell Cleare
BILL NUMBER: S2916 SPONSOR: CLEARE PURPOSE: The bill will study the appropriateness of Hudson Valley counties' placement in the school aid formula regional cost index. SUMMARY OF PROVISIONS: Section one of the bill directs the commissioner of education and the commissioner of labor to study the appropriateness of county placement in the regional cost index used for the school aid formula.