Bill S2875 version original - Leroy Comrie
BILL NUMBER: S2875 SPONSOR: COMRIE TITLE OF BILL: An act to amend the elder law, in relation to establishing a senior dental services grant program PURPOSE OR GENERAL IDEA OF BILL: This bill would establish a grant program for the provision of dental services exclusively for seniors in need of assistance. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends the Elder Law by adding a new section 215-c.