Bill S2940 version original - Pamela Helming
BILL NUMBER: S2940 SPONSOR: HELMING TITLE OF BILL: An act to amend the correction law, in relation to contraband screening at correctional facilities PURPOSE: This bill would require the Commission of the Department of Corrections and Community Supervision (DOCCS) to establish a contraband screening plan in correctional facilities. SUMMARY OF PROVISIONS: Section 1: Section 112 of the Correction Law is amended to increase security and restrict contraband from entering the correction facility.