Bill S3007 version original - Cordell Cleare
BILL NUMBER: S3007 SPONSOR: CLEARE TITLE OF BILL: An act to amend the executive law, in relation to establishing the office for an age-friendly New York PURPOSE: This bill will establish an office within the executive dedicated to making New York a safer and healthier place for older adults to live. This Office for an Age-Friendly New York (OANY) will provide essential services and support to aging individuals and their families.