Bill S2986 version original - Brian Kavanagh
BILL NUMBER: S2986 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the insurance law and the state finance law, in relation to providing for state assistance to local governments for the enforce- ment of fire prevention and building codes PURPOSE: To ensure funding of local code enforcement activities. SUMMARY OF PROVISIONS: Section 1 amends Section 9108 of the state insurance law addressing the fire insurance fee.