Bill S2999 version original - Andrew Gounardes
BILL NUMBER: S2999 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the election law, in relation to requiring the state board of elections to fit registration forms to standard paper size PURPOSE OR GENERAL IDEA OF BILL: The purpose of this legislation is to ensure voter registration forms are fit to standard paper size. SUMMARY OF PROVISIONS: Section 1 amends subdivision 5 of section 5-210 to specify voter regis- tration forms must be sized to standard 8.5 x 11 inches paper.