Bill S3000 version original - Andrew Gounardes
BILL NUMBER: S3000 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the education law, in relation to the registration of nurse anesthetists PURPOSE: To codify current nurse anesthetist practice as the rendering of anes- thesia services under medical supervision. SUMMARY OF SPECIFIC PROVISIONS: Section 1 -- Adds a subdivision 4 to section 6902 of the Education Law.