Bill S3022 version original - Leroy Comrie
BILL NUMBER: S3022 SPONSOR: COMRIE TITLE OF BILL: An act to amend the banking law, in relation to the mortgage bill of rights pamphlet on residential mortgages PURPOSE OR GENERAL IDEA OF BILL: Requires information to consumers of the mortgage bill of rights that are provided by mortgage lenders and bankers to be published in the eight most. spoken languages in New York State. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Adding a new section 35 to the banking law.