Bill S3030 version original - Dan Stec
BILL NUMBER: S3030 SPONSOR: STEC TITLE OF BILL: CONCURRENT RESOLUTION OF THE SENATE AND ASSEMBLY proposing an amendment to section 1 of article 14 of the Constitution, in relation to conveying land to Debar Pond Institute Inc. in order to facilitate the preserva- tion of historic buildings PURPOSE: Amends the constitution to convey land to Debar Pond Institute in order to facilitate the preservation of historic buildings.