Bill S3004 version original - Cordell Cleare
BILL NUMBER: S3004 SPONSOR: CLEARE TITLE OF BILL: An act to amend the elder law, in relation to creating the office of older adult workforce development within the office for the aging PURPOSE: This bill establishes an Office of Older Adult Workforce Development within the State Office for the Aging. SUMMARY OF PROVISIONS: Section one of the bill adds new section 226 to the Elder Law, to estab- lish an office of older adult workforce development within the State office for the Aging.