Bill S3036A version original - Samra Brouk
BILL NUMBER: S3036 SPONSOR: BROUK TITLE OF BILL: An act to amend the education law and the social services law, in relation to the practice of school psychology PURPOSE: This bill provides for licensure of school psychologists and defines the educational and experiential qualifications necessary for licensure and continued practice. SUMMARY OF SPECIFIC PROVISIONS: Section one provides the legislative intent. Section two amends the education law by adding a new Article 166-A.