Bill S3035 version original - Dan Stec
BILL NUMBER: S3035 SPONSOR: STEC TITLE OF BILL: An act to amend the correction law, in relation to confinement; and to repeal certain provisions of such law relating thereto PURPOSE: To repeal the Humane Alternatives to Long-Term Solitary Confinement (HALT) Act enacted in Chapter 93 of the Laws of 2021 and subsequent amendments. SUMMARY OF PROVISIONS: Section 1 amends subdivision 23 of section 2 of the correction law to reflect the original language prior to the enactment of the HALT Act.