Bill S3031 version original - Cordell Cleare
BILL NUMBER: S3031 SPONSOR: CLEARE TITLE OF BILL: An act to amend the civil practice law and rules, in relation to inter- est to be recovered upon a sum awarded in certain circumstances PURPOSE OF THE BILL: To bring actions for personal injury in line with other actions for purposes of computation of interest on a judgment. SUMMARY OF PROVISIONS: Section one amends the CPLR to require calculation of interest in cases for personal injury to be measured from the date the cause of action existed.