Bill S3021 version original - Leroy Comrie
BILL NUMBER: S3021 SPONSOR: COMRIE TITLE OF BILL: An act to amend the energy law, in relation to the composition of the state energy planning board PURPOSE: The bill will restore consumer representation on the State Energy Plan- ning Board. SUMMARY OF PROVISIONS: Amends subdivision 1 of section 6-102 of the energy law, directing the Secretary of State to appoint a representative from the Consumer Protection Division's Utility Intervention Unit to serve on the State Energy Planning Board.