Bill S2959 version original - Brian Kavanagh
BILL NUMBER: S2959 SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the real property tax law, in relation to assessments of residential cooperative, condominium and rental property PURPOSE: This bill would clarify Section 581 of the real property tax law in relation to the assessment of real property owned or leased by a co-op or on a condominium basis. SUMMARY OF PROVISIONS: Section one of the bill creates a new subdivision five in Section 581 of the real property tax law.