Bill S3081 version original - Andrew J. Lanza
BILL NUMBER: S3081 SPONSOR: LANZA TITLE OF BILL: An act to amend the education law, in relation to pre-employment inves- tigations for personnel employed in school districts PURPOSE: This legislation requires the investigation of school district personnel involved with the care, custody or supervision of a child. SUMMARY OF PROVISIONS: Section 1. The education law is amended by adding a new section 3001-e. Section 2.