Bill S3108A version original - John Mannion
BILL NUMBER: S3108 SPONSOR: MANNION TITLE OF BILL: An act to amend the mental hygiene law, in relation to establishing the independent developmental disability ombudsman program PURPOSE: The purpose of this bill is to create the independent developmental disability ombudsman program. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of this bill would add a new section to the Mental Hygiene Law to create the independent developmental disabilities ombudsman program.