Bill S3084 version original - Andrew J. Lanza
BILL NUMBER: S3084 SPONSOR: LANZA PURPOSE: Allows the Town of Massena to issue bonds not to exceed $5 million to cover the cost of the pension contributions due by the town for employ- ees of the former Massena Memorial Hospital. SUMMARY OF PROVISIONS: Section 1 - details the amount and provisions related to the bonding authority of the Town of Massena. Section 2 - marks the effective date. JUSTIFICATION: The Town of Massena recently transferred its municipal hospital to St.