Bill S3082 version original - Andrew J. Lanza
BILL NUMBER: S3082 SPONSOR: LANZA TITLE OF BILL: An act to amend the public service law, in relation to utility rates charged to not-for-profit organized sports programs for youth PURPOSE: Requires utilities to charge not-for-profit organized sports programs for youth residential rates SUMMARY OF PROVISIONS: Section 1. Section 76 of the public service law, as amended by chapter 82 of the laws of 1998, is amended by including not-for-profit organized sports programs for youth. Section 2.