Bill S3073 version original - Andrew J. Lanza
BILL NUMBER: S3073 SPONSOR: LANZA TITLE OF BILL: An act to amend the tax law, in relation to authorizing a personal income tax deduction for elementary and secondary school teachers for certain expenses incurred for school supplies PURPOSE: To provide a personal income tax deduction for school supplies paid for out-of-pocket by K-12 teachers in public and non-public schools, up to $500 a year. SUMMARY OF PROVISIONS: Section 1.