Bill S3067 version original - Dan Stec
BILL NUMBER: S3067 SPONSOR: STEC TITLE OF BILL: An act to amend the public authorities law, in relation to establishing a commission on hosting the olympic games PURPOSE: An act to amend the public authorities law, in relation to establishing a commission on hosting the Olympic games. SUMMARY OF PROVISIONS: Section 1 amends the public authorities law to create a commission on hosting the Olympic games.