Bill S3065B version original - Jessica Ramos
BILL NUMBER: S3065 SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law, in relation to requiring training to reduce abusive conduct and bullying, and cyberbullying in the workplace PURPOSE OR GENERAL IDEA OF BILL: To require all state employees to receive training to prevent abusive conduct and bullying in all state agencies SUMMARY OF PROVISIONS: Section 1: Legislative intent. Section 2.