Bill S3109 version original - John Mannion
BILL NUMBER: S3109 SPONSOR: MANNION TITLE OF BILL: An act to amend the state law, in relation to naming the wood frog the official amphibian of the state of New York PURPOSE: The purpose of this bill is to name the wood frog as the official state amphibian. SUMMARY OF PROVISIONS: Section 1 of this bill would amend the State Law to name the wood frog (lithobates sylvaticus) the official amphibian of the state of New York.