Bill S3099 version original - James Sanders Jr.
BILL NUMBER: S3099 SPONSOR: SANDERS TITLE OF BILL: An act to amend the general municipal law, in relation to requiring individuals or corporations to obtain a sound permit from a municipality in certain situations PURPOSE: This legislation would require individuals or corporations to obtain a sound permit from a municipality in certain situations. SUMMARY OF PROVISIONS: Section 1. Requires individuals or corporations to obtain a sound permit from a municipality in certain situations.