Bill S3056 version original - Pamela Helming
BILL NUMBER: S3056 SPONSOR: HELMING TITLE OF BILL: An act to amend the education law, in relation to requiring a service center for independent living to be located in the county of Ontario PURPOSE: To include in the list of service centers for independent living eligi- ble for Education Department funding. SUMMARY OF PROVISIONS: Amends section 1124 of the education law to add the County of Ontario as a service center for independent living.