Bill S3080 version original - Andrew J. Lanza
BILL NUMBER: S3080 SPONSOR: LANZA TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to sanctions for driving while ability impaired while holding a conditional license PURPOSE: Requires judicial discretion for the issuance of a conditional license. SUMMARY OF PROVISIONS: Section 1. provides for judicial consent at the time of issuance of a conditional license. Section 2.