Bill S3040 version original - Dan Stec
BILL NUMBER: S3040 SPONSOR: STEC TITLE OF BILL: An act to amend the penal law, in relation to exempting any land located in the state forest preserve in the Adirondack and Catskill parks from certain provisions of law relating to prohibiting firearms in public parks PURPOSE: To exempt the Adirondack and Catskill forest preserves from certain provisions of Chapter 371 of 2022. SUMMARY OF PROVISIONS: Section 1 - Exempts the forest preserves in the Adirondack and Catskill parks from 265.