Bill S3043 version original - Zellnor Myrie
BILL NUMBER: S3043 SPONSOR: MYRIE TITLE OF BILL: An act to amend the election law, in relation to eliminating the requirements for publication of detailed election notices within the city of New York PURPOSE: To eliminate detailed publication requirements for election notices. SUMMARY OF PROVISIONS: Section 1 amends subdivision 1 of section 4-118 of the election law which eliminates the requirement for the city of New York in relation to posting primary elections notices.