Bill S3046 version original - Zellnor Myrie
BILL NUMBER: S3046 Revised 1/23/24 SPONSOR: MYRIE TITLE OF BILL: An act to amend the criminal procedure law, in relation to eligibility for conviction sealing for certain applicants PURPOSE: To amend statute 160.59 of the criminal procedure law, and expand the accessibility and effectiveness of conviction sealing made available by the 2017 Raise the Age legislation. SUMMARY OF PROVISIONS: Section 1 amends statute 160.