Bill S3079 version original - Andrew J. Lanza
BILL NUMBER: S3079 SPONSOR: LANZA TITLE OF BILL: An act to amend the general business law, in relation to price gouging PURPOSE OR GENERAL IDEA OF BILL: To further define "unconscionably excessive price" in the context of price gouging. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends Section 396-r of the General Business law. The bill provides that during periods of abnormal disruption of the market for consumer goods and services vital and necessary for the health, safety and welfare.