Bill S3090 version original - Andrew J. Lanza
BILL NUMBER: S3090 SPONSOR: LANZA TITLE OF BILL: An act to amend the penal law and the criminal procedure law, in relation to promoting sexual performances by a child PURPOSE: Elevates promoting an obscene sexual performance by a child and promot- ing a sexual performance by a child to be class B felonies. SUMMARY OF PROVISIONS: Section one amends the closing paragraph of section 263.10 of the penal law and Section two amends the closing paragraph of section 263.