Bill S3088 version original - Andrew J. Lanza
BILL NUMBER: S3088 SPONSOR: LANZA PURPOSE OR GENERAL IDEA OF BILL: This bill amends Education Law section 3604 subdivision 8 to prohibit school on either primary or general election day in a city with a popu- lation of one million or more SUMMARY OF PROVISIONS: Section 1: Prohibits school on primary or general election day in a city with one million or more residents. Section 2: Provides for the Act to take effect immediately.