Bill S3127 version original - John Mannion
BILL NUMBER: S3127 SPONSOR: MANNION TITLE OF BILL: An act to amend the vehicle and traffic law and the navigation law, in relation to implementing oral fluid tests PURPOSE: The purpose of this bill is to require the implementation and use of cannabis field sobriety oral fluids tests. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of this bill would amend subdivision 1 of section 1194 of the Vehicle and Traffic Law by adding a new paragraph.