Bill S2473A version A - Toby Ann Stavisky
BILL NUMBER: S2473A SPONSOR: STAVISKY TITLE OF BILL: An act to amend the business corporation law, the partnership law and the limited liability company law, in relation to certified public accountants PURPOSE OR GENERAL IDEA OF BILL: This bill would amend the education law to allow non-CPAs to be minority owners of CPA firms, consistent with 49 other states and jurisdictions. SUMMARY OF SPECIFIC PROVISIONS: Section 1.