Bill S3178 version original - Cordell Cleare
BILL NUMBER: S3178 SPONSOR: CLEARE TITLE OF BILL: An act to amend the public authorities law, in relation to requiring fare parity for paratransit services operated by certain authorities PURPOSE: To ensure that Access-A-Ride passengers are given fare parity with respect to other users of the system. SUMMARY OF PROVISIONS: Ensures that Access-A-Ride passengers are given fare parity with respect to other users of the system.