Bill S3170C version original - James Skoufis
BILL NUMBER: S3170 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the domestic relations law and the family court act, in relation to establishing "Kyra's Law" PURPOSE OR GENERAL IDEA OF BILL: To protect children by ensuring courts assess any risk to their life and safety in child custody and visitation proceedings. SUMMARY OF PROVISIONS: Section 1 of the bill identifies the name of this legislation as "Kyra's Law.