Bill S3344C version original - James Skoufis
BILL NUMBER: S3344 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the public health law, in relation to requiring patient transporters to meet certain minimum requirements PURPOSE: Requires patient transporters to have training in Basic Life Support. SUMMARY OF PROVISIONS: Section 1: Creates a new section 2830 of the public health law, which sets a definition for patient transporter, and sets a requirement that a patient transporter must be trained in Basic Life Support.