Bill S3264 version original - Brad Hoylman-Sigal
BILL NUMBER: S3264 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the general obligations law, in relation to prohibiting agreements between employers that directly restrict the current or future employment of any employee PURPOSE OF BILL: This bill would prohibit "no-poaching" agreements between certain employers. SUMMARY OF PROVISIONS OF BILL: Section 1 of the bill created a short title of the bill, to be known as the "End Employer Collusion Act".