Bill S3171 version original - Cordell Cleare
BILL NUMBER: S3171 SPONSOR: CLEARE TITLE OF BILL: An act to amend the penal law, in relation to sex trafficking of a vulnerable person PURPOSE OR GENERAL IDEA OF BILL: This bill would make it easier for prosecutors to bring charges against those who exploit disabled individuals for sex trafficking, and would establish the crime of predatory sex trafficking as a class A-II felony. SUMMARY OF PROVISIONS: Section 1 amends the penal law by adding a new subdivision 2 to § 230.