Bill S3285A version original - Kevin Thomas
BILL NUMBER: S3285 SPONSOR: THOMAS TITLE OF BILL: An act to amend the general business law, in relation to children's non-regulated camps PURPOSE: The purpose of this legislation is to require all unregulated children's camps in the state to register with the department of state. SUMMARY OF PROVISIONS: Section 1. Amends section 398-f of the general business law by adding new subdivisions 4, 5, 6, and 7 to require all unregulated children's camps to register with the Department of State.