Bill S3163 version original - Brad Hoylman-Sigal
BILL NUMBER: S3163 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the general business law, in relation to restricting the disclosure of personal information by businesses SUMMARY OF PROVISIONS: Section One states that this Act shall be known and cited as the "Right to Know Act of 2018." Section Two states the legislative intent. Section Three changes the article heading of article 39-F of the General Business.