Bill S3345 version original - James Skoufis
BILL NUMBER: S3345 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the general municipal law, in relation to limiting the naming of municipalities and subdivisions thereof PURPOSE: To prohibit municipalities from having a name containing the word swast- ika and requires municipalities which has a name containing the word. swastika in it to change such name as to remove the word swastika.