Bill S3332 version original - Shelley Mayer
BILL NUMBER: S3332 SPONSOR: MAYER TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to making certain exceptions for violations pertaining to certain mufflers and exhaust systems PURPOSE: To require individuals that receive noise violations as a result of a modified muffler to appear before a judge and cannot receive an adminis- trative dismissal. SUMMARY OF PROVISIONS: Section 1 adds violations originating from noise to the list of non-cor- rectable equipment violations.