Bill S3263 version original - Brad Hoylman-Sigal
BILL NUMBER: S3263 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the civil practice law and rules, in relation to vacat- ing arbitration awards on the basis of arbitrator disregard of the law PURPOSE: To permit an arbitration award to be vacated as a result of an arbitrator's manifest disregard of the law. SUMMARY OF SPECIFIC PROVISIONS: This bill would add a new subdivision to CPLR 7511, which would permit an arbitration award to be vacated due to arbitral manifest disregard of law.