Bill S3334 version original - Shelley Mayer
BILL NUMBER: S3334 SPONSOR: MAYER TITLE OF BILL: An act to amend the executive law, in relation to standards for the erection of tents or membrane structures PURPOSE OR GENERAL IDEA OF BILL: To allow municipalities to extend the use period of temporary tents or membrane structures for the purpose of outdoor dining. SUMMARY OF PROVISIONS: Section 1 specifies that the State Fire Prevention and Building Code shall include provision for municipalities to permit outdoor dining tents year round.