Bill S3173 version original - Cordell Cleare
BILL NUMBER: S3173 SPONSOR: CLEARE TITLE OF BILL: An act to amend the general business law, in relation to the prohibition of the sale of infant inclined sleepers and the restriction of the use of such sleepers in certain settings PURPOSE OR GENERAL IDEA OF BILL: To ban the sale of infant inclined sleepers and restrict their use in certain settings. SUMMARY OF SPECIFIC PROVISIONS: Section 1 adds a new section to the general business law, section 399-iii.