Bill S3238 version original - Luis R. Sepúlveda
BILL NUMBER: S3238 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the labor law and the public health law, in relation to requiring small businesses to have and maintain a first aid kit on the premises of the business at all times PURPOSE OR GENERAL IDEA OF BILL: This bill would require first aid kits to be on premises and available at all small businesses. The existence of such kits would be checked during core enforcements currently done.