Bill S3339 version original - Shelley Mayer
BILL NUMBER: S3339 SPONSOR: MAYER TITLE OF BILL: An act to amend the social services law, in relation to citizen review panels PURPOSE: To expand and regionalize the citizen review panels that oversee child protective services cases. SUMMARY OF PROVISIONS: Section 1. Amends Subdivision 1 of section 371-b of the social services law to establish one citizens review panel in each of the ten economic development regions in the state instead of the minimum of three currently in law.