Bill S3236 version original - Luis R. Sepúlveda
BILL NUMBER: S3236 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the executive law, in relation to defining a victim of a sexual offense PURPOSE: To amend the definition of "victims of a sexual offense" to include victims of revenge porn. SUMMARY OF PROVISIONS: Section 1 amends the definition of a "victim of a sexual offense" to mean a victim of any act constitution an offense as defined under arti- cle one hundred thirty, and/or sections 245.15, 255.25. 255.26, and 255.