Bill S3239 version original - Luis R. Sepúlveda
BILL NUMBER: S3239 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the penal law, in relation to criminal mischief in the third degree and criminal mischief in the second degree SUMMARY OF PROVISIONS: Section 1 - Amends Section 145.05 of the penal law, as amended by chap- ter 276 of the laws of 2003. Section 2 - Amends Section 145.10 of the penal law, as amended by chap- ter 961 of the laws of 1971. Section 3 - establishes an effective date.