Bill S3291 version original - Zellnor Myrie
BILL NUMBER: S3291 SPONSOR: MYRIE TITLE OF BILL: An act to amend the election law, in relation to the date on which a county committee shall meet PURPOSE: To amend the election law committee shall meet in relation to the date on which county committees meet. SUMMARY OF PROVISIONS: Section 1 amends paragraph (b) of subdivision 1 of section 2-112 of the election law which changes the earliest day a county committee can meet from seventeen days to one day after the June primary is certified.