Bill S3169 version original - Cordell Cleare
BILL NUMBER: S3169 SPONSOR: CLEARE TITLE OF BILL: An act to amend the administrative code of the city of New York, in relation to requiring alternating high and low, two-toned signal devices on emergency services vehicles PURPOSE OR GENERAL IDEA OF BILL: To address one of the most common NYC noise complaints-excessively loud sirens. SUMMARY OF SPECIFIC PROVISIONS: Amends the NYC Administrative Code to give clearer guidance on the type of emergency sirens that are permissible.