Bill S3246 version original - Neil D. Breslin
BILL NUMBER: S3246 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the not-for-profit corporation law, in relation to land banks PURPOSE: This legislation allows land banks to make loans and grants. SUMMARY OF PROVISIONS: Section 1: Adds new clauses (22) to § 1607(a) of the Not-For-Profit Corporation Law. It authorizes land banks to make loans and grants. It also allows them to hold real and personal property as security for loans. Section 2: Establishes the effective date.