Bill S3240 version original - Luis R. Sepúlveda
BILL NUMBER: S3240 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the correction law, in relation to the definition of a sex offense PURPOSE OR GENERAL IDEA OF BILL: This bill clarifies when a person convicted of luring a child pursuant to section 120.70 of the penal law is required to register as a sex offender. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of amends the correction law to clarify that a person convicted of luring a child pursuant to section 120.