Bill S3172 version original - Cordell Cleare
BILL NUMBER: S3172 SPONSOR: CLEARE TITLE OF BILL: An act to amend the general business law, in relation to prohibiting the sale of infant loungers and restricting the use of such infant loungers in certain settings PURPOSE OR GENERAL IDEA OF BILL: To ban the sale of infant loungers and restrict their use in certain settings. SUMMARY OF SPECIFIC PROVISIONS: Section 1 adds a new section to the general business law, section 399-iii.