Bill S3230A version original - Brad Hoylman-Sigal
BILL NUMBER: S3230 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the public health law, in relation to adding "embryo" to the definition of "tissue" for purposes of tissue storage SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends subdivision 10 of Section 4360 of the public health law to add "embryo" to the definition of tissues covered by the tissue bank and storage facility licensing law. Section 2 establishes the effective date.