Bill S3309A version original - James Skoufis
BILL NUMBER: S3309 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to Gold Star families PURPOSE: The purpose of this bill is to allow all veterans and members of Gold Star families have free access to public campsites, parks, and other places of recreation in this state. SUMMARY OF PROVISIONS: Section (1) amends section 13.