Bill S3203 version original - Pamela Helming
BILL NUMBER: S3203 SPONSOR: HELMING TITLE OF BILL: An act to amend the state finance law and the executive law, in relation to enacting the "American made flag act" PURPOSE: To ensure that all American flags purchased for use by the State of New York are manufactured in the United States of America. SUMMARY OF PROVISIONS: Sections 1: Titles the bill the "American made flag act.