Bill S513B version A - Kevin Thomas
BILL NUMBER: S513A SPONSOR: THOMAS TITLE OF BILL: An act to amend the education law, in relation to establishing certain school holidays PURPOSE: The purpose of the bill is to update school holiday policies so that students from different religious and cultural backgrounds are able to celebrate with their family and community. SUMMARY OF PROVISIONS: Section 1 of the bill amends section 3210 of the education law section and adds a new section 3204-a.