Bill S3277 version original - Joseph A. Griffo
BILL NUMBER: S3277 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the public officers law, in relation to disclosure related to appointments by the governor and senate PURPOSE: This bill would amend the public officers law by requiring the Governor, Temporary President of the Senate and the Senate Minority Leader to develop a questionnaire to be filled out and completed by any applicant for appointment to office by the governor. SUMMARY OF PROVISIONS: Section 1.