Bill S3300 version original - Brad Hoylman-Sigal
BILL NUMBER: S3300 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to repeal subdivision 4 of section 308 of the county law, relat- ing to access to records of calls made to a municipality's E911 system PURPOSE:: Subdivision 4 of section 308 of the County Law prohibits the disclosure of records from enhanced 911 (or "E911") emergency systems. This bill repeals that prohibition to make such records subject to the Freedom of Information Law (FOIL).