Bill S3192 version original - Jessica Ramos
BILL NUMBER: S3192 SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law, in relation to establishing the unemploy- ment bridge program; to amend the state finance law, in relation to establishing the unemployment bridge program fund; and making an appro- priation therefor (Part A); and to amend the tax law, in relation to enacting the "Digital Ad Tax Act (DATA)" (Part B) SUMMARY OF PROVISIONS: Section 1. Establishes the act to, be cited as the "The Unemployment Bridge Program." Section 2.