Bill S3152 version original - Anthony H. Palumbo
BILL NUMBER: S3152 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the penal law, in relation to criminal possession and sale of stimulants weighing fifteen grams or more PURPOSE: The purpose of this bill is to make the possession and/or sale of high quantities of methamphetamine a class A-1 felony. SUMMARY OF PROVISIONS: Section 220.