Bill S3276 version original - Joseph A. Griffo
BILL NUMBER: S3276 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the penal law, in relation to establishing the offense of aggravated rape and criminal sentencing for defendants convicted of certain sexual offenses PURPOSE: The purpose of this bill is to create an indefinite sentence for indi- viduals convicted of certain sexual offenses. SUMMARY OF PROVISIONS: Section 1 of the bill sets forth the title.