Bill S3154 version original - Anthony H. Palumbo
BILL NUMBER: S3154 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the penal law, in relation to assault on a straphanger PURPOSE: This bill would make it a C felony to push someone onto the subway tracks. SUMMARY OF PROVISIONS: The penal law is amended by adding a new section 120.80 which reads: A person is guilty of assault on a straphanger when he or she strikes, shoves, kicks or otherwise subjects another person to physical contact that results in such person falling onto subway tracks.