Bill S3150 version original - Anthony H. Palumbo
BILL NUMBER: S3150 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the criminal procedure law, in relation to allowing judges more options for when to impose bail or commit the principal to the custody of the sheriff PURPOSE: To allow judges more options for when to impose bail or committing to the custody of the sheriff. SUMMARY OF PROVISIONS: Section 1. Places burglary in the second degree as defined in subdivi- sion two of section 140.