Bill S3260 version original - Brad Hoylman-Sigal
BILL NUMBER: S3260 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the estates, powers and trusts law, in relation to disqualification as a surviving spouse PURPOSE: The purpose of the bill is to permit the disqualification of a spouse to take a share of a decedent's estate if there is an annulment of the marriage before or after the death of the decedent. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Section 5-1.