Bill S3261 version original - Brad Hoylman-Sigal
BILL NUMBER: S3261 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to repeal section 470 of the judiciary law, relating to allowing attorneys having offices in the state to reside in an adjoining state PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to remove the current inability of lawyers admitted in New York to practice in the State if they neither reside nor have an office in the state. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill repeals section 470 of the Judiciary law.