Bill S3341 version original - James Skoufis
BILL NUMBER: S3341 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the public health law, in relation to increasing the age required to use ultraviolet radiation devices to twenty-one years of age. PURPOSE: To reduce the risk of skin cancer by raising the minimum age to use UV tanning beds from 18 to 21 years old. SUMMARY OF PROVISIONS: Section 1: Amends subdivision 2 of section 3555 of the public health law to raise the minimum age to use tanning beds from 18 to 21 years.