Bill S3179 version original - Brad Hoylman-Sigal
BILL NUMBER: S3179 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the education law, in relation to providing for absence from school for the mental or behavioral health of the minor SUMMARY OF PROVISIONS: Section 1 of the bill amends Section 3210 of the Education Law to provide that the absence of a minor from school due to mental or behav- ioral health shall be permitted under rules that the Commissioner of Education shall establish. It also makes technical changes to existing law.