Bill S3408 version original - Kevin S. Parker
BILL NUMBER: S3408 SPONSOR: PARKER TITLE OF BILL: An act to amend the education law, in relation to establishing the bilingual teachers of tomorrow teacher recruitment and retention program PURPOSE OR GENERAL IDEA OF BILL: To set forth requirements and guidelines for the Bilingual Teachers of Tomorrow Recruitment and Retention Program. SUMMARY OF SPECIFIC PROVISIONS: Section 1.