Bill S3417 version original - Zellnor Myrie
BILL NUMBER: S3417 SPONSOR: MYRIE TITLE OF BILL: An act to amend the criminal procedure law, in relation to eligibility for conviction sealing for certain applicants PURPOSE OR GENERAL IDEA OF BILL:: To amend statute 160.59 of the criminal procedure law, and expand the accessibility and effectiveness of conviction sealing. SUMMARY OF SPECIFIC PROVISIONS:: Section 1 amends section 160.