Bill S3384 version original - Timothy M. Kennedy
BILL NUMBER: S3384 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the highway law, in relation to authorizing and direct- ing the department of transportation to establish a memorial marker program PURPOSE: To establish a formal program in New York to allow for the placement of memorials at the location of fatal highway accidents on state highways and roads. SUMMARY OF PROVISIONS: Section 1 adds a new section 67 to highway law. Section 2 establishes the effective date.