Bill S3387 version original - Liz Krueger
BILL NUMBER: S3387 SPONSOR: KRUEGER TITLE OF BILL: An act to amend the executive law, in relation to enacting the online revenues and expenditures transparency act PURPOSE: Directs the Director of Office of Information Technology and the Direc- tor of the Division or the Budget to create and maintain one-stop-shop budget database web site for all of the most up-to-date State spending and revenue statistics. SUMMARY OF PROVISIONS: Section 1.